mt8163:frameworks/av:add required changes for mt8163

 Conflicts:
	media/libaudiohal/2.0/Android.bp
	media/libaudiohal/2.0/DeviceHalLocal.cpp
	media/libmedia/Android.bp
	media/libmedia/MediaProfiles.cpp
	media/libmedia/include/media/MediaProfiles.h
	media/libmediaextractor/Android.bp
	media/libmediaextractor/include/media/stagefright/MediaBufferGroup.h
	media/libstagefright/Android.bp
	services/audioflinger/Android.mk
	services/audioflinger/Threads.cpp
	services/audiopolicy/service/AudioPolicyService.cpp
	services/camera/libcameraservice/Android.mk
	services/camera/libcameraservice/api1/CameraClient.cpp

+++ media/libmedia/IOMX.cpp
@@ -520,8 +520,8 @@ status_t BnOMXNode::onTransact(
             CHECK_OMX_INTERFACE(IOMXNode, data, reply);
 
             OMX_INDEXTYPE index = static_cast<OMX_INDEXTYPE>(data.readInt32());
-
-            size_t size = data.readInt64();
+//test
+            size_t size = data.readInt32();
 
             status_t err = NOT_ENOUGH_DATA;
             void *params = NULL;
